+++ cpp/AlgoLibR/ai/nlp/mining/word/ngram_counter.cpp
@@ -52,7 +52,7 @@ void Node::RemoveChild(const wchar_t key){
 
 
 
-NGramCounter::NGramCounter(const unsigned long min_n, const unsigned long max_n, const wchar_t *delimiters) : min_n(min_n), max_n(max_n), min_freq(0){
+NGramCounter::NGramCounter(const uint8_t min_n, const uint8_t max_n, const wchar_t *delimiters) : min_n(min_n), max_n(max_n), min_freq(0){
     if(delimiters == nullptr) return;
     
     auto n_delim = std::wcslen(delimiters);
@@ -83,7 +83,7 @@ Node* NGramCounter::AddNGram(const wchar_t gram[]){
 std::vector<std::wstring> NGramCounter::ParseLine(const std::wstring &line){
     std::vector<std::wstring> grams;
     // std::wstring gram;
-    unsigned long k = 0;
+    uint8_t k = 0;
     bool is_ok = true;
     for(auto i = 0; i < line.size()-this->max_n; i++){
         is_ok = true;
